move Cadastro e Login para pasta Auth
