--- src/pages/login.css
@@ -1,376 +1,0 @@
-body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: rgb(2, 50, 4);
-  background: linear-gradient(
-    265deg,
-    rgba(2, 50, 4, 1) 0%,
-    rgba(0, 7, 0, 1) 95%,
-    rgba(0, 4, 0, 1) 100%
-  );
-}
-
-/* body {
-    background: #020513ee;
-    font-family: "Montserrat", sans-serif;
-  } */
-
-.container-login {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 120%;
-  margin-bottom: 10rem;
-  height: 25rem;
-  /* background-image: linear-gradient(285deg, #313868 8%, #111728 100%); */
-  cursor: pointer;
-  color: black;
-}
-.div-lubi:hover {
-  transition: 1s;
-  background-image: linear-gradient(225deg, #818dd8 5%, #01050f 100%);
-  transform: translateY(-1rem);
-}
-
-.wrappr-login {
-  display: flex;
-  flex-direction: column;
-  padding: 40px 40px;
-  height: 19.7rem;
-  background-image: linear-gradient(225deg, #3043bd 5%, #01050f 100%);
-  border-radius: 4px;
-  width: 34.8rem;
-  z-index: 10;
-  cursor: pointer;
-}
-
-.div-lubi::before {
-  content: "";
-  position: absolute;
-  top: -60%;
-  left: -10%;
-  width: 390px;
-  height: 440px;
-  background: linear-gradient(0deg, transparent, yellow, yellow);
-  transform-origin: bottom right;
-  animation: animate 6s linear infinite;
-}
-
-.div-lubi::after {
-  content: "";
-  position: absolute;
-  top: -60%;
-  left: -10%;
-  width: 390px;
-  height: 440px;
-  background: linear-gradient(0deg, transparent, #43f3ff, #45f3ff);
-  transform-origin: bottom right;
-  animation: animate 6s linear infinite;
-  animation-delay: -3s;
-}
-
-@keyframes animate {
-  0% {
-    transform: rotate(0deg);
-  }
-
-  100% {
-    transform: rotate(360deg);
-  }
-}
-.div-lubi {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 25rem;
-  border-radius: 4px;
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.5);
-  width: 40rem;
-  margin: 2rem auto;
-  cursor: pointer;
-  overflow: hidden;
-}
-
-.wrappr-login p {
-  display: flex;
-  margin-top: 10px;
-  color: #00a8fb;
-  font-size: 12pt;
-}
-
-.digite {
-  display: flex;
-
-  width: 100%;
-  height: 1.5rem;
-  align-items: center;
-  justify-content: center;
-  font-family: "Special Elite", monospace;
-}
-
-.input-wrappre {
-  position: relative;
-  width: 100%;
-  margin-top: 2em;
-}
-.input-wrappre input {
-  position: relative;
-  width: 100%;
-  padding: 20px 10px 10px;
-  background: transparent;
-  border: none;
-  outline: none;
-  font-size: 1em;
-  letter-spacing: 0.05em;
-  color: white;
-  z-index: 10;
-}
-.input-wrappre span {
-  position: absolute;
-  left: 0;
-  top: 0;
-  padding: 20px 10px 10px;
-  font-size: 0.9em;
-  transition: 0.7s;
-  color: white;
-}
-
-.input-wrappre input:valid ~ span,
-.input-wrappre input:focus ~ span {
-  color: #00a8fb;
-  font-family: "Special Elite", monospace;
-  transform: translateX(-10px) translateY(-35px);
-  font-size: 0.75em;
-}
-.input-wrappre i {
-  position: absolute;
-  left: 0;
-  bottom: 0px;
-  width: 100%;
-  height: 2px;
-  border-radius: 4px;
-  transition: 0.7s;
-  pointer-events: none;
-  background-color: #43f3ff;
-  z-index: 9;
-}
-
-.input-wrappre input:valid ~ i,
-.input-wrappre input:focus ~ i {
-  height: 45px;
-}
-
-.mybotton-login {
-  display: flex;
-  margin-top: 120px;
-
-  width: 100%;
-  height: auto;
-
-  align-items: center;
-  justify-content: space-around;
-
-  position: relative;
-}
-
-.loginlogar {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 3px;
-  border-radius: 10px;
-  position: relative;
-  cursor: pointer;
-}
-.loginlogar:hover {
-  transition: 0.9s;
-  transform: translateX(20px);
-}
-.loginlogar::after,
-.loginlogar::before {
-  content: "";
-  position: absolute;
-  bottom: 0;
-  top: 0;
-  right: 0;
-  left: 0;
-  transition: transform 0.3s ease-in-out;
-  border-radius: 5px;
-}
-
-.loginlogar::before {
-  transform: scaleY(0);
-  border-right: 2px solid #09f315;
-  border-left: 2px solid #f51d09;
-}
-.loginlogar::after {
-  transform: scaleX(0);
-  border-top: 2px solid #fbbc05;
-  border-bottom: 2px solid #4285f4;
-}
-.loginlogar:hover:before {
-  transform: scaleX(1);
-}
-.loginlogar:hover::after {
-  transform: scaleX(1);
-}
-
-/**=================stilos de botão 1========**/
-.a-dobotaoa {
-  position: relative;
-  display: inline-block;
-  padding: 15px 30px;
-  color: #2196f3;
-  text-decoration: none;
-  text-transform: uppercase;
-  letter-spacing: 5px;
-  font-size: 25px;
-  overflow: hidden;
-  transition: 0.5s;
-}
-.a-dobotaoa:hover {
-  color: #255784;
-  background-color: #2196f3;
-  box-shadow: 0 0 10px #2196f3, 0 0 40px #2196f3, 0 0 80px #2196f3;
-  transition-delay: 1s;
-}
-
-.spanbotao {
-  position: absolute;
-  display: block;
-}
-
-.spanbotao:nth-child(1) {
-  top: 0;
-  left: -100%;
-  width: 100%;
-  height: 2px;
-  background: linear-gradient(90deg, transparent, #09f315 10%, #4285f4 50%);
-}
-
-.a-dobotaoa:hover .spanbotao:nth-child(1) {
-  left: 100%;
-  transition: 1s;
-}
-
-.spanbotao:nth-child(3) {
-  bottom: 0;
-  right: -100%;
-  width: 100%;
-  height: 2px;
-  background: linear-gradient(270deg, transparent, pink 1%, #fbbc05 50%);
-}
-
-.a-dobotaoa:hover .spanbotao:nth-child(3) {
-  right: 100%;
-  transition: 1s;
-  transition-delay: 0.5s;
-}
-
-.spanbotao:nth-child(2) {
-  top: -100%;
-  right: 0;
-  width: 2px;
-  height: 100%;
-  background: linear-gradient(180deg, transparent, #09f315);
-}
-
-.a-dobotaoa:hover .spanbotao:nth-child(2) {
-  top: 100%;
-  transition: 1s;
-  transition-delay: 0.25s;
-}
-
-.spanbotao:nth-child(4) {
-  bottom: -100%;
-  left: 0;
-  width: 2px;
-  height: 100%;
-  background: linear-gradient(360deg, transparent, #f51d09);
-}
-
-.a-dobotaoa:hover .spanbotao:nth-child(4) {
-  bottom: 100%;
-  transition: 1s;
-  transition-delay: 0.75s;
-}
-
-/**=================stilos de botão 2========**/
-
-.linkeparacadastro {
-  position: relative;
-  display: inline-block;
-  padding: 5px 15px;
-  color: #2196f3;
-  font-size: 15px;
-  overflow: hidden;
-  transition: 1s;
-}
-.linkeparacadastro:hover {
-  color: #255784;
-  background-color: #2196f3;
-  box-shadow: 0 0 10px #2196f3, 0 0 40px #2196f3, 0 0 80px #2196f3;
-  transition-delay: 1s;
-}
-
-.spanlinke {
-  position: absolute;
-  display: block;
-}
-
-.spanlinke:nth-child(1) {
-  top: 0;
-  left: -100%;
-  width: 100%;
-  height: 2px;
-  background: linear-gradient(90deg, transparent, #4285f4);
-}
-
-.linkeparacadastro:hover .spanlinke:nth-child(1) {
-  left: 100%;
-  transition: 1s;
-}
-
-.spanlinke:nth-child(3) {
-  bottom: 0;
-  right: -100%;
-  width: 100%;
-  height: 2px;
-  background: linear-gradient(270deg, transparent, #fbbc05);
-}
-
-.linkeparacadastro:hover .spanlinke:nth-child(3) {
-  right: 100%;
-  transition: 1s;
-  transition-delay: 0.5s;
-}
-
-.spanlinke:nth-child(2) {
-  top: -100%;
-  right: 0;
-  width: 2px;
-  height: 100%;
-  background: linear-gradient(180deg, transparent, #09f315);
-}
-
-.linkeparacadastro:hover .spanlinke:nth-child(2) {
-  top: 100%;
-  transition: 1s;
-  transition-delay: 0.25s;
-}
-
-.spanlinke:nth-child(4) {
-  bottom: -100%;
-  left: 0;
-  width: 2px;
-  height: 100%;
-  background: linear-gradient(360deg, transparent, #f51d09);
-}
-
-.linkeparacadastro:hover .spanlinke:nth-child(4) {
-  bottom: 100%;
-  transition: 1s;
-  transition-delay: 0.75s;
-}
